requirements.txt
asgiref==3.2.7
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.41.0
django-cors-headers==3.2.1
django-embed-video==1.3.2
django-rest-auth==0.9.5
djangorestframework==3.11.0
flake8==3.8.2
idna==2.9
importlib-metadata==1.6.0
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.1.2
pycodestyle==2.6.0
pyflakes==2.2.0
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
stripe==2.48.0
urllib3==1.25.9
zipp==3.1.0
